+++ R/run_sttest.R
@@ -17,11 +17,11 @@ run_sttest <- function() {
             raw$base[raw$base[,'statGroups'] %in% stats$STTest.Pairs[i,"g2"],]$Value ))
         if (stats$STTest.Pairs$ftest[[i]][[1]]$p.value > 0.05) {
             stats$STTest.Pairs$ftest[[i]][2] = "Homogenous"
-            message(sprintf("variances between group1 %s & group2 %s ARE homogenous with a p-value: %s", stats$STTest.Pairs[i,"g1"], stats$STTest.Pairs[i,"g2"], stats$STTest.Pairs$ftest[[i]][[1]]$p.value))
+            histova_msg(sprintf("variances between group1 %s & group2 %s ARE homogenous with a p-value: %s", stats$STTest.Pairs[i,"g1"], stats$STTest.Pairs[i,"g2"], stats$STTest.Pairs$ftest[[i]][[1]]$p.value))
         } else {
             stats$STTest.Pairs$ftest[[i]][2] = "NOT Homogenous - CHECK DATA"
-            message(sprintf("** CHECK DATA - variances between group1 %s & group2 %s ARE NOT homogenous with a p-value: %s", stats$STTest.Pairs[i,"g1"], stats$STTest.Pairs[i,"g2"], stats$STTest.Pairs$ftest[[i]][[1]]$p.value))
-            warning(sprintf("** CHECK DATA - variances between group1 %s & group2 %s ARE NOT homogenous with a p-value: %s (file: %s)", stats$STTest.Pairs[i,"g1"], stats$STTest.Pairs[i,"g2"], stats$STTest.Pairs$ftest[[i]][[1]]$p.value, the$Location.File))
+            histova_msg(sprintf("** CHECK DATA - variances between group1 %s & group2 %s ARE NOT homogenous with a p-value: %s", stats$STTest.Pairs[i,"g1"], stats$STTest.Pairs[i,"g2"], stats$STTest.Pairs$ftest[[i]][[1]]$p.value))
+            histova_msg(sprintf("** CHECK DATA - variances between group1 %s & group2 %s ARE NOT homogenous with a p-value: %s (file: %s)", stats$STTest.Pairs[i,"g1"], stats$STTest.Pairs[i,"g2"], stats$STTest.Pairs$ftest[[i]][[1]]$p.value, the$Location.File), type="warn")
         }
         variance = TRUE
         if (stats$STTest.Pairs$var[i] == "unequal") variance = FALSE
@@ -34,7 +34,7 @@ run_sttest <- function() {
             var.equal = variance,
             paired = paired,
             conf.level=0.95)
-        message(sprintf("ran a %s students t-test on group1 %s to group2 %s, with tail: %s, variance: %s, p-value: %s", toupper(stats$STTest.Pairs[i,"pair"]), stats$STTest.Pairs[i,"g1"], stats$STTest.Pairs[i,"g2"], toupper(stats$STTest.Pairs[i,"alt"]), toupper(stats$STTest.Pairs[i,"var"]), stats$STTest.Pairs$sttest[[i]]$p.value))
+        histova_msg(sprintf("ran a %s students t-test on group1 %s to group2 %s, with tail: %s, variance: %s, p-value: %s", toupper(stats$STTest.Pairs[i,"pair"]), stats$STTest.Pairs[i,"g1"], stats$STTest.Pairs[i,"g2"], toupper(stats$STTest.Pairs[i,"alt"]), toupper(stats$STTest.Pairs[i,"var"]), stats$STTest.Pairs$sttest[[i]]$p.value))
     } # for (i in rownames(Stats.STTest.Pairs)) {
     #assign("Stats.STTest.Pairs", Stats.STTest.Pairs, envir = .GlobalEnv) ### CHANGED - no longer needed ###
 }