+++ R/run_wttest.R
@@ -7,7 +7,7 @@
 run_wttest <- function() {
     # old code from the wilcox t-test, need to add in again as w-test or something...
     if (FALSE) {
-        message("---- Paired Wilcox T-Test")
+        histova_msg("---- Paired Wilcox T-Test")
         for (i in rownames(stats$PTTest.Pairs)) {
             i = as.numeric(i)
 
@@ -21,7 +21,7 @@ run_wttest <- function() {
                 var.equal = variance,
                 paired=TRUE,
                 conf.level=0.95)
-            message(sprintf("ran paired wilcox t-test on group1 %s to group2 %s, with tail: %s, variance: %s, p-value: %s", stats$PTTest.Pairs[i,"g1"], stats$PTTest.Pairs[i,"g2"], stats$PTTest.Pairs[i,"alt"], stats$PTTest.Pairs[i,"var"], stats$PTTest.Pairs$pttest[[i]]$p.value))
+            histova_msg(sprintf("ran paired wilcox t-test on group1 %s to group2 %s, with tail: %s, variance: %s, p-value: %s", stats$PTTest.Pairs[i,"g1"], stats$PTTest.Pairs[i,"g2"], stats$PTTest.Pairs[i,"alt"], stats$PTTest.Pairs[i,"var"], stats$PTTest.Pairs$pttest[[i]]$p.value))
         }
         #assign("Stats.PTTest.Pairs", Stats.PTTest.Pairs, envir = .GlobalEnv) ### CHANGED - not needed ###
     }